+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * github-260: Remove hive-storage-api from several plugins and lib
 * github-261: Add descriptions to all pom.xml
 * github-262: Verification of "relation" targets should only check existence
+* github-263: Add filter condition to data quality checks in documentation
 
 
 # Version 0.28.0 - 2022-10-07